Bulk-changing many long to int32_t
Done with:
        sed -i -e 's/unsigned long/uint32_t/g' *.cpp *.h
        sed -i -e 's/ulong/uint32_t/g' *.cpp *.h
        sed -i -e 's/signed long/int32_t/g' *.cpp *.h
        sed -i -e 's/long/int32_t/g' *.cpp *.h
        sed -i -e 's/uint32_t int32_t/unsigned long long/g' *.cpp *.h
        sed -i -e 's/int32_t int32_t/long long/g' *.cpp *.h

+++ src/TTGenerator.h
@@ -59,7 +59,7 @@ public:
 	virtual void Call(short actParams, short anyNum[], short functNum);
 	virtual void Asm(bool inLine, wchar_t text[], wchar_t error[]);
 	virtual void Quit(void);
-	virtual void InitTTGenerator(TrueTypeFont *font, TrueTypeGlyph *glyph, long glyphIndex, TTEngine *tt, bool legacyCompile, bool *memError);
+	virtual void InitTTGenerator(TrueTypeFont *font, TrueTypeGlyph *glyph, int32_t glyphIndex, TTEngine *tt, bool legacyCompile, bool *memError);
 	virtual void TermTTGenerator(void);
 	TTGenerator(void);
 	virtual ~TTGenerator(void);
